third_party/WebKit/LayoutTests/fast/dom/Window/dispatchEvent-expected.txt

Test window.dispatchEvent().

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS window.dispatchEvent(event) threw exception InvalidStateError: An attempt was made to use an object that is not, or is no longer, usable..
PASS window.dispatchEvent(event) threw exception InvalidStateError: An attempt was made to use an object that is not, or is no longer, usable..
PASS myEventDispatched is true
PASS target is window
PASS currentTarget is window
PASS useCaptureDispatched is true
PASS nonUseCaptureDispatched is true
PASS successfullyParsed is true

TEST COMPLETE
